Move all classes related to base android components into androidcomponents package
