+++ app/src/main/java/com/epam/themes/androidcomponents/SimpleFragment.java
@@ -1,9 +1,8 @@
-package com.epam;
+package com.epam.themes.androidcomponents;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.support.annotation.WorkerThread;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
