+++ app/src/main/java/com/epam/themes/androidcomponents/activities/AActivity.java
@@ -1,4 +1,4 @@
-package com.epam.activities;
+package com.epam.themes.androidcomponents.activities;
 
 import android.os.Bundle;
