+++ app/src/main/java/com/epam/themes/androidcomponents/activities/BaseActivity.java
@@ -1,4 +1,4 @@
-package com.epam.activities;
+package com.epam.themes.androidcomponents.activities;
 
 import android.app.Activity;
 import android.content.Intent;
